+++ src/tezos/tezosEngine.js
@@ -2,6 +2,7 @@
 import { bns } from 'biggystring'
 import {
   type EdgeCurrencyEngineOptions,
+  type EdgeFetchFunction,
   type EdgeSpendInfo,
   type EdgeTransaction,
   type EdgeWalletInfo,
@@ -15,17 +16,16 @@ import {
   asyncWaterfall,
   getOtherParams,
   makeMutex,
-  promiseAny,
-  validateObject
+  promiseAny
 } from '../common/utils.js'
 import { TezosPlugin } from '../tezos/tezosPlugin.js'
 import { currencyInfo } from './tezosInfo.js'
-import { XtzTransactionSchema } from './tezosSchema.js'
 import {
   type HeadInfo,
   type OperationsContainer,
   type TezosOperation,
-  type XtzGetTransaction
+  type XtzGetTransaction,
+  asXtzGetTransaction
 } from './tezosTypes.js'
 
 const ADDRESS_POLL_MILLISECONDS = 15000
@@ -46,14 +46,17 @@ type TezosFunction =
 
 export class TezosEngine extends CurrencyEngine {
   tezosPlugin: TezosPlugin
+  fetchCors: EdgeFetchFunction
 
   constructor(
     currencyPlugin: TezosPlugin,
     walletInfo: EdgeWalletInfo,
-    opts: EdgeCurrencyEngineOptions
+    opts: EdgeCurrencyEngineOptions,
+    fetchCors: EdgeFetchFunction
   ) {
     super(currencyPlugin, walletInfo, opts)
     this.tezosPlugin = currencyPlugin
+    this.fetchCors = fetchCors
   }
 
   async multicastServers(func: TezosFunction, ...params: any): Promise<any> {
@@ -93,15 +96,14 @@ export class TezosEngine extends CurrencyEngine {
 
       case 'getNumberOfOperations':
         funcs = this.tezosPlugin.tezosApiServers.map(server => async () => {
-          const result = await this.io
-            .fetch(
-              `${server}/v3/number_operations/${params[0]}?type=Transaction`
-            )
+          const result = await this.fetchCors(
+            `${server}/v1/accounts/${params[0]}`
+          )
             .then(function (response) {
               return response.json()
             })
             .then(function (json) {
-              return json[0]
+              return json.numTransactions
             })
           return { server, result }
         })
@@ -110,17 +112,15 @@ export class TezosEngine extends CurrencyEngine {
 
       case 'getTransactions':
         funcs = this.tezosPlugin.tezosApiServers.map(server => async () => {
-          const pagination = /mystique/.test(server)
+          const pagination = /tzkt/.test(server)
             ? ''
             : `&p='${params[1]}&number=50`
-          const result: XtzGetTransaction = await this.io
-            .fetch(
-              `${server}/v3/operations/${params[0]}?type=Transaction` +
-                pagination
-            )
-            .then(function (response) {
-              return response.json()
-            })
+          const result: XtzGetTransaction = await this.fetchCors(
+            `${server}/v1/accounts/${params[0]}/operations?type=transaction` +
+              pagination
+          ).then(function (response) {
+            return response.json()
+          })
           return { server, result }
         })
         out = await asyncWaterfall(funcs)
@@ -221,22 +221,20 @@ export class TezosEngine extends CurrencyEngine {
   }
 
   processTezosTransaction(tx: XtzGetTransaction) {
-    const valid = validateObject(tx, XtzTransactionSchema)
-    if (!valid) {
-      this.log('Invalid transaction!')
-      throw new Error('InvalidTransactionError')
-    }
+    const transaction = asXtzGetTransaction(tx)
     const pkh = this.walletLocalData.publicKey
     const ourReceiveAddresses: Array<string> = []
     const currencyCode = PRIMARY_CURRENCY
-    const date = new Date(tx.type.operations[0].timestamp).getTime() / 1000
-    const blockHeight = tx.type.operations[0].op_level
-    let nativeAmount = tx.type.operations[0].amount.toString()
-    const networkFee = tx.type.operations[0].fee.toString()
-    const failedOperation = tx.type.operations[0].failed
-    if (pkh === tx.type.operations[0].destination.tz) {
+    const date = new Date(transaction.timestamp).getTime() / 1000
+    const blockHeight = transaction.level
+    let nativeAmount = transaction.amount.toString()
+    const networkFee = (
+      transaction.bakerFee + transaction.allocationFee
+    ).toString()
+    const failedOperation = transaction.status === 'failed'
+    if (pkh === transaction.target.address) {
       ourReceiveAddresses.push(pkh)
-      if (tx.type.source.tz === pkh) {
+      if (transaction.sender.address === pkh) {
         nativeAmount = '-' + networkFee
       }
     } else {
@@ -359,11 +357,13 @@ export class TezosEngine extends CurrencyEngine {
     await this.startEngine()
   }
 
-  async makeSpend(edgeSpendInfoIn: EdgeSpendInfo) {
+  async makeSpend(edgeSpendInfoIn: EdgeSpendInfo): Promise<EdgeTransaction> {
     return makeSpendMutex(() => this.makeSpendInner(edgeSpendInfoIn))
   }
 
-  async makeSpendInner(edgeSpendInfoIn: EdgeSpendInfo) {
+  async makeSpendInner(
+    edgeSpendInfoIn: EdgeSpendInfo
+  ): Promise<EdgeTransaction> {
     const {
       edgeSpendInfo,
       currencyCode,