+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # edge-currency-accountbased
 
+# 0.7.17 (2020-08-04)
+
+- FIO checkTransactions algorithm update to page transactions
+- Fix REPv2 token address
+
+# 0.7.16 (2020-07-29)
+
+- Add REPV2 ERC20 token
+
+# 0.7.15 (2020-07-23)
+
+- Add new Tezos API
+- FIO - fix multicastServers
+
 # 0.7.14 (2020-07-12)
 
 - FIO fix domain reg url